Backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
streamlit